feat(#6): CharacterMovement Script, Colliders for the Character, Slippery Material, Slide Input
added move and slide into the character movement script, this now has to get the right animations. for slide there will be 2 animations depending on the ground the character is on. add a box & circle collider and also ground and ceiling detection to the character. add slippery material so the character cant be stuck on edges.